palona_ai_sales_system/requirements.txt
agency-swarm>=0.1.0
openai>=1.0.0
python-dotenv>=0.19.0
streamlit>=1.30.0
requests>=2.26.0
python-docx>=1.0.0
PyPDF2>=3.0.0
numpy>=1.21.0
pydantic>=1.8.2
python-dateutil>=2.8.2
pytz>=2021.3
pandas>=1.3.0
beautifulsoup4>=4.9.3
lxml>=4.9.0
google-search-results>=2.4.2
aiohttp>=3.8.0 